move config.Init into config package
